+++ propulsion-projector/Handler.fs
@@ -1,10 +1,28 @@
 module ProjectorTemplate.Handler
 
-#if cosmos
+//#if cosmos
 #if     parallelOnly
 // Here we pass the items directly through to the handler without parsing them
 let mapToStreamItems (x : System.Collections.Generic.IReadOnlyList<'a>) : seq<'a> = upcast x
 #else // cosmos && !parallelOnly
+#if    synthesizeSequence // cosmos && !parallelOnly && !synthesizeSequence
+let indices = Propulsion.Kafka.StreamNameSequenceGenerator()
+
+let parseDocumentAsEvent (doc : Microsoft.Azure.Documents.Document) : Propulsion.Streams.StreamEvent<byte[]> =
+    let docId = doc.Id
+    //let streamName = Propulsion.Streams.StreamName.internalParseSafe docId // if we're not sure there is a `-` in the id, this helper adds one
+    let streamName = FsCodec.StreamName.parse docId // throws if there's no `-` in the id
+    let ts = let raw = doc.Timestamp in raw.ToUniversalTime() |> System.DateTimeOffset
+    let docType = "DocumentTypeA" // each Event requires an EventType - enables the handler to route without having to parse the Data first
+    let data = string doc |> System.Text.Encoding.UTF8.GetBytes
+    // Ideally, we'd extract a monotonically incrementing index/version from the source and use that
+    // (Using this technique neuters the deduplication mechanism)
+    let streamIndex = indices.GenerateIndex streamName
+    { stream = streamName; event = FsCodec.Core.TimelineEvent.Create(streamIndex, docType, data, timestamp=ts) }
+
+let mapToStreamItems (docs : Microsoft.Azure.Documents.Document seq) : Propulsion.Streams.StreamEvent<byte[]> seq =
+    docs |> Seq.map parseDocumentAsEvent
+#else // cosmos && !parallelOnly && synthesizeSequence
 //let replaceLongDataWithNull (x : FsCodec.ITimelineEvent<byte[]>) : FsCodec.ITimelineEvent<_> =
 //    if x.Data.Length < 900_000 then x
 //    else FsCodec.Core.TimelineEvent.Create(x.Index, x.EventType, null, x.Meta, timestamp=x.Timestamp)
@@ -14,12 +32,13 @@ let mapToStreamItems (x : System.Collections.Generic.IReadOnlyList<'a>) : seq<'a
 
 let mapToStreamItems (docs : Microsoft.Azure.Documents.Document seq) : Propulsion.Streams.StreamEvent<_> seq =
     docs
-    |> Seq.collect  Propulsion.Cosmos.EquinoxCosmosParser.enumStreamEvents
+    |> Seq.collect Propulsion.Cosmos.EquinoxCosmosParser.enumStreamEvents
     // TODO use Seq.filter and/or Seq.map to adjust what's being sent etc
     // |> Seq.map hackDropBigBodies
+#endif // cosmos && !parallelOnly && synthesizeSequence
 #endif // !parallelOnly
-#endif // cosmos
-//#if esdb
+//#endif // cosmos
+#if esdb
 open Propulsion.EventStore
 
 /// Responsible for inspecting and then either dropping or tweaking events coming from EventStore
@@ -28,7 +47,7 @@ let tryMapEvent filterByStreamName (x : EventStore.ClientAPI.ResolvedEvent) =
     match x.Event with
     | e when not e.IsJson || e.EventStreamId.StartsWith "$" || not (filterByStreamName e.EventStreamId) -> None
     | PropulsionStreamEvent e -> Some e
-//#endif // esdb
+#endif // esdb
 
 #if kafka
 #if     (cosmos && parallelOnly) // kafka && cosmos && parallelOnly